06_week/project/live_dashboard/requirements.txt
python-socketio==5.2.1
eventlet==0.30.2